+++ trio/_core/_exceptions.py
@@ -123,3 +123,12 @@ class BrokenResourceError(Exception):
     information about the underlying error.
 
     """
+
+
+class EndOfChannel(Exception):
+    """Raised when trying to receive from a :class:`trio.abc.ReceiveChannel`
+    that has no more data to receive.
+
+    This is analogous to an "end-of-file" condition, but for channels.
+
+    """