+++ trio/tests/test_highlevel_serve_listeners.py
@@ -14,7 +14,7 @@ from trio.testing import memory_stream_pair, wait_all_tasks_blocked
 class MemoryListener(trio.abc.Listener):
     closed = attr.ib(default=False)
     accepted_streams = attr.ib(factory=list)
-    queued_streams = attr.ib(factory=(lambda: trio.open_channel(1)))
+    queued_streams = attr.ib(factory=(lambda: trio.open_memory_channel(1)))
     accept_hook = attr.ib(default=None)
 
     async def connect(self):